+++ CHANGES.rst
@@ -6,6 +6,12 @@ Changelog
 - new PATCH version for backwards compatible bug fixes
 
 
+0.1.3
+-----
+2020-07-17: feature release
+    - comprehensive *--colortest*
+    - automatically select 8 colors profile for travis
+
 0.1.2
 -----
 2020-07-16: feature release